Fix indexing bug in value assignment for r_addr.
r_addr was being set to NULL under the wrong conditions (including
when the buffer was not fully parsed because of a new line).
Additionally, buf was incorrectly conceived of as equivalent to
str when it is only the last READ_SIZE bytes of  str or less,
causing the incorrect value to be assigned to r_addr. This commit
addresses these mistakes.