Use c executable instead of python script for process to hijack.
For the time being this allows me to more precisely know what I
can expect to be in the heap and what I can expect instead to be
in the stack. I.e.
I can ensure that the string I want to replace is in the heap if
I use malloc to create the string in C.